+++ src/worker/core/GameSim.football/index.ts
@@ -2178,6 +2178,7 @@ class GameSim {
 						clock: this.clock,
 						t,
 						names: [p.name],
+						injuredPID: p.id,
 					});
 				}
 			}